requirements.txt
auto-py-to-exe
bcrypt
beautifulsoup4
browser-cookie3
colorama
cryptography
requests
deep-translator
discord
dnspython
exifread
customtkinter
tkinter
pyautogui
pycryptodome
pyinstaller
pyqt5
pyqtwebengine
pywin32
pyzipper
rarfile
screeninfo
selenium
setuptools
urllib3
whois
whoiam